Please read this note completely!!!
This is the child branch.

I changed the structure a little bit.

A main change is that I treat SixesWildApplication as "theGame" instead
of SixesWild.

You should consider change HashMap back to 2-D array because we don't
need the extra fitures HashMap brings to us.

But the current version is using HashMap and Position class as the key
class. The previous implementation of HashMap in master branch is
totally wrong.

Please look really carefully at my code. Especially TileLabelController
and NormalSelectionMove.

I created an interface IMove for convience. Either IMove or abstract
Move class are ok for me.

The board is playable in a sense you can select and remove squares. But
the refilling of squares are not 100% done yet.

The current structure of Controller, Boundary and Entities(including
moves) are correct. We can develope other controllers and moves in this
structure.